backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
SQLAlchemy==2.0.23
sqlite3
python-dotenv==1.0.0
email-validator==2.1.0
pydantic[email]==2.5.0
aiofiles==23.2.1